ATR-8: fixed bug in the code fix verifier.
The verifier previously used only the first diagnostic in the set to get available code actions. If that diagnostic couldn't provide any code action to apply, then the verifier stopped. This led to a situation when the first diagnostic in the code source didn't provide an action, and the verifier stops even if the second diagnostic can provide valid code actions